requirements.txt
black==24.8.0
cookiecutter==2.6.0
crispy-tailwind==1.0.3
django-browser-reload==1.15.0
django-environ==0.11.2
django-extensions==3.2.3
django-htmx==1.19.0
django-tailwind==3.8.0
djhtml==3.0.6
docopt==0.6.2
humanfriendly==10.0
ipython==8.12.3
nbconvert==7.16.4
pipdeptree==2.23.1
setuptools==58.1.0
yarg==0.1.9
